+++ other/joisc2007/src/problemI_circuit.cpp
@@ -43,7 +43,7 @@ int main() {
 		if (v[s] > 0) {
 			dp[0].clear();
 			dp[0][0] = true;
-			for (unsigned long int i = 0; i < d.size(); i++) {
+			for (int i = 0; i < int(d.size()); i++) {
 				bool f = 1 == gcd(s, k / d[i]);
 				for (int j = 0; j <= v[s]; j++) {
 					dp[i + 1][j] = dp[i][j];
